updates with-google-tag-manager example to use script component
